Refactor client-generator templates and tests
- Updated import paths in test files to reflect new structure.
- Created new template files for API response types, configuration structure, deserializer types, form utilities, operation utilities, parameter serialization, response parsing, and template utilities.
- Implemented rendering functions for API response parsing utilities and types, configuration interfaces, and default values.
- Added utility functions for handling FormData and URL encoding.
- Introduced serialization utilities for OpenAPI parameters.
- Enhanced response parsing capabilities with support for various content types.
- Established a centralized index file for configuration templates to streamline imports and exports.
