+++ packages/client-generator/src/templates/config/template-utils.ts
@@ -1,0 +1,9 @@
+/* Shared template utilities */
+
+/*
+ * Renders Zod import statement
+ */
+export function renderZodImportStatement(): string {
+  return `import type { z } from "zod/v4";
+`;
+}
